requirements/requirements.txt
argparse
requests
paramiko
jenkins-job-builder
anymarkup
configparser